+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Userlist for Ruby Changelog
 
+## Unreleased (main)
+
+- Updates ActiveJob Worker to retry on Timeout::Error with polynomially longer wait times, up to 10 attempts
+
 ## v0.9.0 (2024-03-19)
 
 - Allows deleteing resources by using other identifiers (like email)